DRY the `insertBlockOrPatternViaBlockAppender` function by extracting it to an `AbstractEditorComponent` class
In this case, inheritance was simpler than composing it, due to the two existing editor classes;

- The `GutenbergEditorComponent` and `SiteEditorComponent` now inherit from this abstract class;
- The class provides a noop `runInCanvas` instance method and lets subclasses override the implementation;
- The abstract class can be used to DRY other functions, too, later.
