+++ test/e2e/lib/gutenberg/gutenberg-editor-component.js
@@ -1,6 +1,6 @@
 import { kebabCase } from 'lodash';
 import webdriver, { By } from 'selenium-webdriver';
-import AsyncBaseContainer from '../async-base-container';
+import AbstractEditorComponent from '../components/abstract-editor-component';
 import GuideComponent from '../components/guide-component.js';
 import * as driverHelper from '../driver-helper';
 import * as driverManager from '../driver-manager.js';
@@ -9,7 +9,7 @@ import { FileBlockComponent } from './blocks/file-block-component';
 import { ImageBlockComponent } from './blocks/image-block-component';
 import { ShortcodeBlockComponent } from './blocks/shortcode-block-component';
 
-export default class GutenbergEditorComponent extends AsyncBaseContainer {
+export default class GutenbergEditorComponent extends AbstractEditorComponent {
 	constructor( driver, url, editorType = 'iframe' ) {
 		super( driver, By.css( '.edit-post-header' ), url );
 		this.editorType = editorType;
@@ -713,24 +713,5 @@ export default class GutenbergEditorComponent extends AsyncBaseContainer {
 		const notices = await this.driver.findElements( locator );
 		await Promise.all( notices.map( ( notice ) => notice.click() ) );
 		await driverHelper.waitUntilElementNotLocated( this.driver, locator );
-	}
-
-	async insertBlockOrPatternViaBlockAppender( name, container = 'Group' ) {
-		const containerBlockId = await this.addBlock( container );
-		const blockAppenderLocator = By.css(
-			`#${ containerBlockId } .block-editor-button-block-appender`
-		);
-		await driverHelper.clickWhenClickable( this.driver, blockAppenderLocator );
-
-		const quickInserterSearchInputLocator = By.css(
-			`.block-editor-inserter__quick-inserter .components-search-control__input`
-		);
-
-		const patternItemLocator = By.css(
-			'.block-editor-inserter__quick-inserter .block-editor-block-types-list__item, .block-editor-inserter__quick-inserter .block-editor-block-patterns-list__item'
-		);
-
-		await driverHelper.setWhenSettable( this.driver, quickInserterSearchInputLocator, name );
-		await driverHelper.clickWhenClickable( this.driver, patternItemLocator );
 	}
 }
